requirements.txt
lyricsgenius>=3.0.0
python-docx>=0.8.11
requests>=2.25.1
PyQt6>=6.0.0
beautifulsoup4>=4.9.0